split catalog into its own file
finish moving catalog. fix exception names
