+++ src/data/__init__.py
@@ -1,3 +1,4 @@
+from .catalog import *
 from .datasets import *
 from .fetch import *
 from .utils import *
